# Microoptimize decompression