@Cyan4973 2020-02-22
I tested this variant, on a recent Intel `x64` system, using `gcc v9.2.1`.
I could measure a small performance improvement from this patch, in the 2-3% range.
That's pretty good !