@Cyan4973 2020-02-22
I tried a variant which would not `force_inline` the `write` functions.

It did not helped to tame the compression speed reduction, which remained identical.
However, it resulted in a large loss of decompression speed, by ~4-5%, large enough to make the new version slower than `dev`. This is all the more strange as the write functions are not used during decompression.

For me, this is a hint that the observed differences may not be related to the code itself, but to indirect (and uncontrollable) effects, such as instruction alignment.

If that's the case, I expect to witness different outcomes by changing compilers and versions.
To be tested.