@Cyan4973 2020-02-22
As feared, changing compiler result in different outcome.

Moving to `clang` `v9.0.0-2`, the decompression speed takes a hit of _almost -10%_ when applying this patch. This is very large.

It could be that `clang` doesn't like the inlining, potentially combined with some other less predictable impacts, such as instruction alignment.