@Cyan4973 2020-02-22
I've updated results for `gcc` and `clang` on `linux x64`.

It looks to me that the outcome of applying this patch is basically random, sometimes better, sometimes worse, on average rather worse.

It's not the fault of the patch, rather it's the complexity of modern cpus, with their strange, changing and mostly undocumented rules when entering the uop cache domain, combined with the inability of compilers to make appropriate choices with regards to instruction alignment. That makes optimization a rather difficult exercise, especially when the impact of the source code is entangled with some random impact from instruction alignment. When the second impact is larger than the first one, it's a real pain.