@wolfpld 2020-02-22
Well, looks like a no-go then.

The code layout affecting performance issue is real, and can be probably analyzed using `coz` (I have only watched the video, so can't say anymore about that):
https://github.com/plasma-umass/coz
https://www.youtube.com/watch?v=r-TLSBdHe1A