@Cyan4973 2020-02-24
Thanks for the links, it's a good watch !
Indeed, we should first develop a way to analyze a code change while neutralizing layout effects, since they tend to be stronger than the change itself.